allow pacstrap to skip mount of pseudo filesystems
It is possible that people embed pacstrap into existing installation
frameworks, such as installimage[0]. Those frameworks may mount all the
needed pseudo filesystems, so pacstrap doesn't need to care about it.
This commit add the -m flag (short for mountfoo). If -m is set, pacstrap
will skip the mount section for pseudofs. The default behaviour is still
to mount them.

[0] https://github.com/virtapi/installimage